Separated keyboard input and tokenizing logic
