--- Assets/EQUIControlsManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 974cb2d44f61b2d4ab3500e022b95856
