+++ Assets/Scenes/BasicScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 65b6a8ec5bf185b4cbb4eb8ddba005e6
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
