+++ Assets/Scripts/GraphManager.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public class GraphManager : MonoBehaviour
+{
+    
+}
